+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to ssql will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [4.94.0] - 2026-09-08
 
 ### Fixed
 - **Aggregation expressions can produce strings, bools and times; `max`/`min`